Move picking jammed towers to generate messages. Add extra jamming functionality. Fix #19.
Generate messages already does a pass over the networks states to decide
what source destinations to pick for messages. Thus it is easier to pick
jammed towers there than repeating the same work inside the simulator.
We now pass an extra list of which towers to be jammed to the moby
simulator.