Add user jamming logic for random and mobility based jamming. Fix #10
Bookkeeping for mobility of users is expensive. Adds a constant time
delay to the generate messages script, but there's no way around this.
The delay isn't too bad, around 3 mins. So it is only trigerred if the
simulation requires it. Bookkeeping for tower jamming changed as well,
to save some cycles.