Normalize achtung job scheduling based on total memory available.
The old method of scheduling jobs involved equally splitting up all jobs
on the number of achtung servers available but this is not the best way
to do it as the achtung servers vary according to ram. Thus, normalizing
the jobs based on how much total ram the machines have is a good idea.